requirements.txt
requests
pyaudio 
openai-whisper 
beautifulsoup4 
sentence-transformers 
faiss-cpu 
numpy 
sounddevice 
pydub 
keyboard 
python-dotenv
gtts       